Resolves #127 - Add thread page - Move thread and comment state to threadSlice - Fix App losing path on first load - Move post component into pages/thread/components
