+++ app/src/pages/thread/Thread.css
@@ -1,0 +1,12 @@
+.Thread-page__content {
+  display: grid;
+  place-items: start center;
+}
+
+@media only screen and (min-width: 920px) {
+  .Thread-page__content {
+    justify-self: center;
+    margin-top: 24px;
+    width: 80%;
+  }
+}
