+++ app/src/pages/thread/components/post/Post.css
