+++ app/src/pages/thread/components/post/index.tsx
